+++ gouge/colourcli.py
@@ -1,4 +1,6 @@
 import logging
+from logging import Handler, LogRecord
+from typing import Any, List, Optional
 
 from blessings import Terminal
 
@@ -17,7 +19,8 @@ class Simple(logging.Formatter):
     """
 
     @staticmethod
-    def basicConfig(show_exc=True, show_threads=False, *args, **kwargs):
+    def basicConfig(show_exc=True, show_threads=False, **kwargs):
+        # type: (bool, bool, Any) -> List[Handler]
         '''
         Convenience method to have a one-liner set-up.
 
@@ -32,24 +35,32 @@ class Simple(logging.Formatter):
         modified. This is useful if you want to modify the handlers any further
         (for example using :py:class:`~gouge.filters.ShiftingFilter`).
         '''
-        logging.basicConfig(*args, **kwargs)
+        logging.basicConfig(**kwargs)
         root = logging.getLogger()
         output = []
         for handler in root.handlers:
-            if (hasattr(handler, 'stream') and
-                    hasattr(handler.stream, 'name') and
-                    handler.stream.name in ('<stderr>', '<stdout>')):
-                handler.setFormatter(Simple(show_exc, show_threads))
-                output.append(handler)
+            stream = getattr(handler, 'stream', None)
+            if not stream:
+                continue
+
+            stream_name = getattr(stream, 'name', None)
+            if stream_name not in ('<stderr>', '<stdout>'):
+                continue
+
+            handler.setFormatter(Simple(show_exc, show_threads))
+            output.append(handler)
         return output
 
-    def __init__(self, show_exc=False, show_threads=False, *args, **kwargs):
-        logging.Formatter.__init__(self, *args, **kwargs)
+    def __init__(self, show_exc=False, show_threads=False, fmt=None,
+                 datefmt=None):
+        # type: (bool, bool, Optional[str], Optional[str]) -> None
+        logging.Formatter.__init__(self, fmt, datefmt)
         self.show_threads = show_threads
         self.show_exc = show_exc
         self.term = Terminal()
 
     def format(self, record):
+        # type: (LogRecord) -> str
         if record.levelno <= logging.DEBUG:
             colorize = self.term.bold_black
         elif record.levelno <= logging.INFO:
@@ -62,7 +73,7 @@ class Simple(logging.Formatter):
             colorize = self.term.bold_yellow_on_red
 
         record.message = record.getMessage()
-        record.asctime = self.formatTime(record, self.datefmt)
+        record.asctime = self.formatTime(record, self.datefmt or '')
 
         message_items = [
             '{t.green}{asctime}{t.normal}',
@@ -79,10 +90,13 @@ class Simple(logging.Formatter):
             if record.exc_info:
                 # Cache the traceback text to avoid converting it multiple times
                 # (it's constant anyway)
-                if not record.exc_text:
-                    record.exc_text = self.formatException(record.exc_info)
+                exc_text = getattr(record, 'exc_text', '')
+                if not exc_text:
+                    record.exc_text = self.formatException(  # type: ignore
+                        record.exc_info)
 
-            if record.exc_text:
+            exc_text = getattr(record, 'exc_text', '')
+            if exc_text:
                 message_template += '\n{t.red}{exc_text}{t.normal}'
 
         return message_template.format(