+++ docs/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+Version 1.2.1
+-------------
+
+* Add type-hints & fix minor typing bugs
+* Add ``gouge.__version__`` attribute
+
 Version 1.2.0
 -------------